Wire up credentialManagers login in the proper flow
