+++ wear/src/main/kotlin/au/com/shiftyjelly/pocketcasts/wear/WearMainActivityViewModel.kt
@@ -1,33 +1,16 @@
 package au.com.shiftyjelly.pocketcasts.wear
 
 import android.content.Context
-import android.util.Log
-import androidx.credentials.CredentialManager
-import androidx.credentials.CustomCredential
-import androidx.credentials.GetCredentialRequest
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import au.com.shiftyjelly.pocketcasts.account.watchsync.WatchSync
-import au.com.shiftyjelly.pocketcasts.account.watchsync.WatchSyncAuthData
 import au.com.shiftyjelly.pocketcasts.models.type.SignInState
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.repositories.refresh.RefreshPodcastsTask
-import au.com.shiftyjelly.pocketcasts.repositories.sync.LoginResult
-import au.com.shiftyjelly.pocketcasts.repositories.sync.SignInSource
-import au.com.shiftyjelly.pocketcasts.repositories.sync.SyncManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
-import au.com.shiftyjelly.pocketcasts.utils.extensions.isGooglePlayServicesAvailableSuccess
-import au.com.shiftyjelly.pocketcasts.utils.log.LogBuffer
-import com.google.android.gms.common.GoogleApiAvailability
-import com.google.android.horologist.auth.data.tokenshare.TokenBundleRepository
-import com.google.android.libraries.identity.googleid.GetSignInWithGoogleOption
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.qualifiers.ActivityContext
 import dagger.hilt.android.qualifiers.ApplicationContext
-import java.util.UUID
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -42,20 +25,15 @@ import timber.log.Timber
 class WearMainActivityViewModel @Inject constructor(
     private val playbackManager: PlaybackManager,
     private val podcastManager: PodcastManager,
-    private val tokenBundleRepository: TokenBundleRepository<WatchSyncAuthData?>,
     private val userManager: UserManager,
     private val settings: Settings,
-    private val watchSync: WatchSync,
     @ApplicationContext private val context: Context,
-    private val syncManager: SyncManager
 ) : ViewModel() {
 
     data class State(
         val showLoggingInScreen: Boolean = false,
         val signInState: SignInState = SignInState.SignedOut,
     )
-
-    private val credentialManager: CredentialManager by lazy { CredentialManager.create(context) }
 
     private val _state = MutableStateFlow(State())
     val state = _state.asStateFlow()
@@ -68,87 +46,6 @@ class WearMainActivityViewModel @Inject constructor(
                 .collect { signInState ->
                     _state.update { it.copy(signInState = signInState) }
                 }
-        }
-    }
-
-    fun tryCredentialsManager(
-        @ActivityContext context: Context,
-    ) {
-        val isGoogleSignInAvailable = Settings.GOOGLE_SIGN_IN_SERVER_CLIENT_ID.isNotEmpty() &&
-            GoogleApiAvailability.getInstance().isGooglePlayServicesAvailableSuccess(context)
-
-        Log.d("===", "tryCredentialsManager ctxt=$context, isgoogleAvail=$isGoogleSignInAvailable")
-
-        if (isGoogleSignInAvailable) {
-            viewModelScope.launch {
-                try {
-                    val googleIdOption: GetSignInWithGoogleOption = GetSignInWithGoogleOption.Builder(
-                        serverClientId = Settings.GOOGLE_SIGN_IN_SERVER_CLIENT_ID,
-                    ).setNonce(UUID.randomUUID().toString()).build()
-
-                    val request: GetCredentialRequest = GetCredentialRequest.Builder()
-                        .addCredentialOption(googleIdOption)
-                        .build()
-
-                    val result = credentialManager.getCredential(
-                        request = request,
-                        context = context,
-                    )
-                    val credential = result.credential as CustomCredential
-                    if (credential.type == GoogleIdTokenCredential.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
-                        val googleIdTokenCredential = GoogleIdTokenCredential
-                            .createFrom(credential.data)
-
-                        signInWithGoogleToken(
-                            idToken = googleIdTokenCredential.idToken,
-                        )
-                    } else {
-                        Log.w("====", "else")
-                        LogBuffer.e(LogBuffer.TAG_INVALID_STATE, "Failed to sign in with Google One Tap")
-                        watchAuthTokenExchange()
-                    }
-                } catch (e: Exception) {
-                    Log.w("====", "error $e")
-                    LogBuffer.e(LogBuffer.TAG_CRASH, e, "Unable to sign in with Google One Tap")
-                    watchAuthTokenExchange()
-                }
-            }
-        } else {
-            watchAuthTokenExchange()
-        }
-    }
-
-    private suspend fun signInWithGoogleToken(
-        idToken: String,
-    ) {
-        val authResult = syncManager.loginWithGoogle(idToken = idToken, signInSource = SignInSource.UserInitiated.Onboarding)
-        onLoginFromPhoneResult(authResult)
-    }
-
-    private fun watchAuthTokenExchange() {
-        viewModelScope.launch {
-            tokenBundleRepository.flow
-                .collect { watchSyncAuthData ->
-                    watchSync.processAuthDataChange(watchSyncAuthData) {
-                        onLoginFromPhoneResult(it)
-                    }
-                }
-        }
-    }
-
-    private fun onLoginFromPhoneResult(loginResult: LoginResult) {
-        when (loginResult) {
-            is LoginResult.Failed -> { /* do nothing */
-            }
-
-            is LoginResult.Success -> {
-                viewModelScope.launch {
-                    podcastManager.refreshPodcastsAfterSignIn()
-                }
-                _state.update {
-                    it.copy(showLoggingInScreen = true)
-                }
-            }
         }
     }
